Book For Everyone/requirements.txt
django-phonenumber-field
phonenumbers
django-crispy-forms
Pillow